correccion repositorios y desarrollo de los servicios

+++ api-cervezas/src/main/java/daw2a/apicervezas/repositories/StyleRepository.java
@@ -3,5 +3,5 @@ package daw2a.apicervezas.repositories;
 import daw2a.apicervezas.entities.Style;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StyleRepository extends JpaRepository<Style, Integer> {
+public interface StyleRepository extends JpaRepository<Style, Long> {
 }
